backend/requirements.txt
# Low Barber - Sistema de Gestión de Barbería
# Dependencias del backend

# Flask y extensiones
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.6.0

# Base de datos
psycopg2-binary==2.9.9

# Serialización y validación
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Utilidades
python-dotenv==1.0.0
Werkzeug==3.0.1

# Generar con: pip freeze > requirements.txt
